Fix [#15] #13에서 변경된 사항 conflict 난 부분 해결 및 리뷰 사항 일부 수정 작업 진행

+++ morib/src/main/java/org/morib/server/api/timerView/dto/TaskInTodoCardDto.java
@@ -10,10 +10,10 @@ public record TaskInTodoCardDto(
     LocalDate endDate,
     boolean isComplete,
     LocalDate targetDate,
-    Long targetTime
+    int targetTime
 ) {
 
-    public static TaskInTodoCardDto of(Task task, LocalDate targetDate, Long targetTime){
+    public static TaskInTodoCardDto of(Task task, LocalDate targetDate, int targetTime){
         return new TaskInTodoCardDto(task.getName(), task.getStartDate(), task.getEndDate(), task.getIsComplete(), targetDate, targetTime);
     }
 }
